tiny bugfix: brainmask uses minimal UNI values
A problem occurs, when no brainmask is provided to voxels of the UNI image with intensity == 0.
As the range of the MP2RAGE image has been already scaled to -0.5 to 0.5, the intensity 0 does not represent the background (as probably intended?). Therefore I suggest to select the minimum values of the current MP2RAGE image instead